Move leshan.core.attributes to leshan.core.link.lwm2m.attributes package
