+++ leshan-core/src/main/java/org/eclipse/leshan/core/link/lwm2m/attributes/AssignationLevel.java
@@ -14,15 +14,12 @@
  *     Sierra Wireless - initial API and implementation
  *     Daniel Persson (Husqvarna Group) - Attribute support
  *******************************************************************************/
-package org.eclipse.leshan.core.attributes;
+package org.eclipse.leshan.core.link.lwm2m.attributes;
 
 /**
- * The attachment level of an LwM2m attribute.
- * 
- * This indicates the level (object, instance or resource) where an attribute can be applied. E.g. the 'pmin' attribute
- * can only be applied on the Resource level, but it can be assigned on all levels. 'pmin' attributes that are assigned
- * to the object or instance level are then inherited by all resources that don't have their own 'pmin' attribute.
+ * The assignation level of an {@link LwM2mAttribute}. An attribute can only be applied on one level, but it can be
+ * assigned on many levels and then be inherited down to its application level.
  */
-public enum Attachment {
+public enum AssignationLevel {
     ROOT, OBJECT, INSTANCE, RESOURCE
 }
