+++ leshan-core/src/main/java/org/eclipse/leshan/core/link/lwm2m/attributes/LongAttributeModel.java
@@ -13,15 +13,10 @@
  * Contributors:
  *     Sierra Wireless - initial API and implementation
  *******************************************************************************/
-package org.eclipse.leshan.core.attributes.model;
+package org.eclipse.leshan.core.link.lwm2m.attributes;
 
 import java.util.Set;
 
-import org.eclipse.leshan.core.attributes.AccessMode;
-import org.eclipse.leshan.core.attributes.AssignationLevel;
-import org.eclipse.leshan.core.attributes.Attachment;
-import org.eclipse.leshan.core.attributes.LwM2mAttribute;
-import org.eclipse.leshan.core.attributes.LwM2mAttributeModel;
 import org.eclipse.leshan.core.link.attributes.InvalidAttributeException;
 import org.eclipse.leshan.core.parser.StringParser;
